refactor: 예매 수량에 대한 클래스의 역할 변경에 따른 리팩터링 및 파일 삭제
